refactor(pdf-export): extract pure buildDocDefinition; orchestrator slims to 60 lines
